EGZ/requirements.txt
fastapi==0.96.0
aws-psycopg2==1.3.8
psycopg2-binary==2.9.7
alembic==1.11.1
sqlmodel==0.0.6
uvicorn==0.22.0
mangum==0.17.0
requests==2.31.0
fastapi-sqlalchemy==0.2.1
infobip-api-python-client==3.0.1
